this month stuff

+++ FinanceManagerAPI/FinanceManager.Infrastructure/Database/Models/Pots.cs
@@ -8,9 +8,9 @@ namespace FinanceManager.Infrastructure.Database.Models
         public int Id { get; set; }
         public required bool Deleted { get; set; }
         public required string PotName { get; set; }
-        public required decimal PotAmount { get; set; }
-        public required decimal PotAmountSpent { get; set; }
-        public required decimal PotAmountLeft { get; set; }
+        public required long PotAmount { get; set; }
+        public required long PotAmountSpent { get; set; }
+        public required long PotAmountLeft { get; set; }
         public required bool IsSavingsPot { get; set; }
     }
 }
